+++ src/oci/apigateway/models/api_specification_route_backend.py
@@ -35,6 +35,10 @@ class ApiSpecificationRouteBackend(object):
     #: This constant has a value of "OAUTH2_LOGOUT_BACKEND"
     TYPE_OAUTH2_LOGOUT_BACKEND = "OAUTH2_LOGOUT_BACKEND"
 
+    #: A constant which can be used with the type property of a ApiSpecificationRouteBackend.
+    #: This constant has a value of "OAUTH2_LOGIN_BACKEND"
+    TYPE_OAUTH2_LOGIN_BACKEND = "OAUTH2_LOGIN_BACKEND"
+
     def __init__(self, **kwargs):
         """
         Initializes a new ApiSpecificationRouteBackend object with values from keyword arguments. This class has the following subclasses and if you are using this class as input
@@ -44,13 +48,14 @@ class ApiSpecificationRouteBackend(object):
         * :class:`~oci.apigateway.models.HTTPBackend`
         * :class:`~oci.apigateway.models.OracleFunctionBackend`
         * :class:`~oci.apigateway.models.StockResponseBackend`
+        * :class:`~oci.apigateway.models.OAuth2LoginBackend`
         * :class:`~oci.apigateway.models.DynamicRoutingBackend`
 
         The following keyword arguments are supported (corresponding to the getters/setters of this class):
 
         :param type:
             The value to assign to the type property of this ApiSpecificationRouteBackend.
-            Allowed values for this property are: "ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND", "OAUTH2_LOGIN_BACKEND", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type type: str
 
@@ -83,6 +88,9 @@ class ApiSpecificationRouteBackend(object):
         if type == 'STOCK_RESPONSE_BACKEND':
             return 'StockResponseBackend'
 
+        if type == 'OAUTH2_LOGIN_BACKEND':
+            return 'OAuth2LoginBackend'
+
         if type == 'DYNAMIC_ROUTING_BACKEND':
             return 'DynamicRoutingBackend'
         else:
@@ -94,7 +102,7 @@ class ApiSpecificationRouteBackend(object):
         **[Required]** Gets the type of this ApiSpecificationRouteBackend.
         Type of the API backend.
 
-        Allowed values for this property are: "ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND", "OAUTH2_LOGIN_BACKEND", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -113,7 +121,7 @@ class ApiSpecificationRouteBackend(object):
         :param type: The type of this ApiSpecificationRouteBackend.
         :type: str
         """
-        allowed_values = ["ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND"]
+        allowed_values = ["ORACLE_FUNCTIONS_BACKEND", "HTTP_BACKEND", "STOCK_RESPONSE_BACKEND", "DYNAMIC_ROUTING_BACKEND", "OAUTH2_LOGOUT_BACKEND", "OAUTH2_LOGIN_BACKEND"]
         if not value_allowed_none_or_none_sentinel(type, allowed_values):
             type = 'UNKNOWN_ENUM_VALUE'
         self._type = type