+++ src/oci/dns/models/zone.py
@@ -457,7 +457,9 @@ class Zone(object):
     def resolution_mode(self):
         """
         **[Required]** Gets the resolution_mode of this Zone.
-        The resolution mode of a zone defines behavior related to how query responses can be handled.
+        The resolution mode of a zone defines behavior related to how query responses can be handled. See `Private DNS Zone Transparency`__ for more information.
+
+        __ https://docs.oracle.com/iaas/Content/DNS/Tasks/privatedns.htm#use-cases__resolution
 
         Allowed values for this property are: "STATIC", "TRANSPARENT", "RTYPE_TRANSPARENT", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
@@ -472,7 +474,9 @@ class Zone(object):
     def resolution_mode(self, resolution_mode):
         """
         Sets the resolution_mode of this Zone.
-        The resolution mode of a zone defines behavior related to how query responses can be handled.
+        The resolution mode of a zone defines behavior related to how query responses can be handled. See `Private DNS Zone Transparency`__ for more information.
+
+        __ https://docs.oracle.com/iaas/Content/DNS/Tasks/privatedns.htm#use-cases__resolution
 
 
         :param resolution_mode: The resolution_mode of this Zone.