+++ src/oci/resource_scheduler/models/work_request.py
@@ -26,6 +26,10 @@ class WorkRequest(object):
     OPERATION_TYPE_STOP_RESOURCE = "STOP_RESOURCE"
 
     #: A constant which can be used with the operation_type property of a WorkRequest.
+    #: This constant has a value of "BACKUP_RESOURCE"
+    OPERATION_TYPE_BACKUP_RESOURCE = "BACKUP_RESOURCE"
+
+    #: A constant which can be used with the operation_type property of a WorkRequest.
     #: This constant has a value of "CHANGE_SCHEDULE_COMPARTMENT"
     OPERATION_TYPE_CHANGE_SCHEDULE_COMPARTMENT = "CHANGE_SCHEDULE_COMPARTMENT"
 
@@ -76,7 +80,7 @@ class WorkRequest(object):
 
         :param operation_type:
             The value to assign to the operation_type property of this WorkRequest.
-            Allowed values for this property are: "START_RESOURCE", "STOP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE", 'UNKNOWN_ENUM_VALUE'.
+            Allowed values for this property are: "START_RESOURCE", "STOP_RESOURCE", "BACKUP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE", 'UNKNOWN_ENUM_VALUE'.
             Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
         :type operation_type: str
 
@@ -160,7 +164,7 @@ class WorkRequest(object):
         **[Required]** Gets the operation_type of this WorkRequest.
         The asynchronous operation tracked by this work request.
 
-        Allowed values for this property are: "START_RESOURCE", "STOP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE", 'UNKNOWN_ENUM_VALUE'.
+        Allowed values for this property are: "START_RESOURCE", "STOP_RESOURCE", "BACKUP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE", 'UNKNOWN_ENUM_VALUE'.
         Any unrecognized values returned by a service will be mapped to 'UNKNOWN_ENUM_VALUE'.
 
 
@@ -179,7 +183,7 @@ class WorkRequest(object):
         :param operation_type: The operation_type of this WorkRequest.
         :type: str
         """
-        allowed_values = ["START_RESOURCE", "STOP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE"]
+        allowed_values = ["START_RESOURCE", "STOP_RESOURCE", "BACKUP_RESOURCE", "CHANGE_SCHEDULE_COMPARTMENT", "CREATE_SCHEDULE", "UPDATE_SCHEDULE"]
         if not value_allowed_none_or_none_sentinel(operation_type, allowed_values):
             operation_type = 'UNKNOWN_ENUM_VALUE'
         self._operation_type = operation_type