+++ src/oci/resource_scheduler/schedule_client_composite_operations.py
@@ -80,7 +80,7 @@ class ScheduleClientCompositeOperations(object):
             __ https://docs.oracle.com/iaas/Content/General/Concepts/identifiers.htm
 
         :param oci.resource_scheduler.models.ChangeScheduleCompartmentDetails change_schedule_compartment_details: (required)
-            This is the change schedule compartment details that will be updated.
+            These are the change schedule compartment details that will be updated.
 
         :param list[str] wait_for_states:
             An array of states to wait on. These should be valid values for :py:attr:`~oci.resource_scheduler.models.WorkRequest.status`